requirements.txt
google-api-python-client==2.21.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
Flask==2.0.1
Flask-WTF==0.15.1
Flask-Babel==2.0.0
Flask-Login==0.5.0
cryptography==3.4.8
paypalhttp==1.0.1
